Fix issue whereby nodes drawn with a composite path could not be moved around the canvas correctly.
Previously, the composite path class was overriding the default behaviour of the move method such that it caused each of its members to be moved. This is problematic because a composite path is represented with a paper.js group, and we need to move the group as well as any paper.js items contained in that group.

The fix is to remove the custom move method in the composite path class and instead allow the default implementation (in the path class) to move the paper.js group. The paper.js library then handles moving the nested elements.