Generate source maps when compiling coffee script.
Note that the version of grunt-contrib-coffee that I'm currently using produces source maps with an incorrect sourceRoot attribute. Upgrading to a later version of grunt-contrib-coffee should fix this, but this seems to necessitate updating to a later version of Node and possibly other dependencies.

For now, a quick workaround was to edit node_modules/grunt-contrib-coffee/tasks/coffee.js at line 133:

     sourceRoot: appendTrailingSlash(path.relative(paths.destDir, path.dirname(file)))

becomes

	sourceRoot: ""