Implement redraw for changes to a label's placement.
As a consequence of adding this feature, the design of labelled paths and stencils was simplified a little. Previously a labelled stencil could return any kind of path when placement was sent to none. Now a labelled stencil always returns a labelled path, and the visibility of the underlying label is controlled by the placement property.